Fix particle spawners not visible since CSM spawner implementation
* Drop the ID mapper, use a big u64 instead. This will permit to resync server ids properly with the manager code
* Modernize some code parts (std::unordered_map, auto)
* generate id on client part on U32_MAX + 1 ids, lower are for server ids

+++ src/client/clientevent.h
@@ -108,7 +108,7 @@ struct ClientEvent
 			u16 attached_id;
 			bool vertical;
 			std::string *texture;
-			u32 id;
+			u64 id;
 			struct TileAnimationParams animation;
 			u8 glow;
 		} add_particlespawner;
